requirements.txt
pymemgpt
pyautogen
openai